restructured the files location
